Move existing geodate related endpoint
